+++ netlify/functions/api.js
@@ -1,9 +1,7 @@
 const express = require("express");
 const serverless = require("serverless-http");
 const loginRouter = require("../../src/RoutePaths/login");
-const homeRouter = require("../../src/RoutePaths/Home");
-const roleRouter = require("../../src/RoutePaths/Role");
-const { verifyIdToken } = require("../../src/authMiddleware");
+const homeRouter = require("../../src/RoutePaths/home");
 
 // Create an instance of the Express app
 const app = express();
@@ -34,11 +32,6 @@ app.use((req, res, next) => {
 
   next();
 })
-
-//For Roles related routes
-app.use("/api/role", roleRouter);
-
-app.use(verifyIdToken);
 
 // For Login routes
 app.use('/api/', loginRouter);
